FLEX-5291 ~ Moves astronomical offset within astronomical time
The description of the astronomical offsets was in a separate section
below astronomical time and astronomical time with light sensor trigger.
The astronomical offset is however only taken into account for
astronomical time (without the light sensor). The description of these
offsets is moved to a subsection of astronomical time to avoid the
confusion that made me document it separately.