Updates IDE set up for formatting conventions
The documentation described how to configure the Eclipse IDE to format
the code in an acceptable way for the XGF code base.

With more developers choosing to use IntelliJ instead of Eclipse,
formatting became an issue with pull requests.
To have guidelines that are more easily supported across different IDEs
the Java formatting conventions are changed to be based on Google Java
Format with some additions to honor earier established conventions that
can be easily enforced in both Eclipse and IntelliJ.

The documentation now points to the Config repository where more detailed
instructions on how to configure the IDEs are available.