Adds description how to add Signed-off-by from Eclipse
Adds a short description with a screen shot of the Eclipse Staging view,
explaining how to add the Signed-off-by line to the Git commit message
from certain versions of Eclipse.

Adds some lines to the command line section to have it stand out more
that this is the lower case '-s' for '--signoff' as well as a comment
that the longer version is needed with "git merge" commands.

Signed-off-by: Bart van der Zwet <bart.van.der.zwet@cgi.com>